Parser updated

+++ LoggingKata/Point.cs
@@ -2,7 +2,7 @@
 {
     public struct Point
     {
-        decimal Longitude { get; set; }
-        decimal Latitude { get; set; }
+        double Longitude { get; set; }
+        double Latitude { get; set; }
     }
 }
